Remove Project struct
